+++ src/account.ts
@@ -1,5 +1,5 @@
-import { createEmitter } from '../utilities'
-import { Emitter, Ac } from '../types'
+import { createEmitter } from './utilities'
+import { Emitter, Ac } from './types'
 import SDK from '.'
 
 function account(
